+++ Cmdlets/Private/ConvertTo-TeamViewerOrganizationalUnit.ps1
@@ -1,0 +1,24 @@
+function ConvertTo-TeamViewerOrganizationalUnit {
+    param(
+        [Parameter(ValueFromPipeline = $true, Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [PSObject]
+        $InputObject
+    )
+
+    process {
+        $properties = @{
+            Id          = $InputObject.id
+            Name        = $InputObject.name
+            Description = $InputObject.description
+            ParentId    = $InputObject.parentId
+            CreatedAt   = $InputObject.createdAt
+            UpdatedAt   = $InputObject.updatedAt
+        }
+
+        $result = New-Object -TypeName PSObject -Property $properties
+        $result.PSObject.TypeNames.Insert(0, 'TeamViewerPS.OrganizationalUnit')
+
+        Write-Output $result
+    }
+}