+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Added
 
+- Adds `Get-TeamViewerOrganizationalUnit` to retrieve the organizational unit details.
+- Adds `New-TeamViewerOrganizationalUnit` to add a new organizational unit.
+- Adds `Remove-TeamViewerOrganizationalUnit` to delete the organizational unit.
+- Adds `Set-TeamViewerOrganizationalUnit` to modify the organizational unit.
 - Adds `Get-TeamViewerCompany` to retrieve the company / tenant details.
 - Adds `Get-TeamViewerLicense` to retrieve company / tenant licenses.
 - Adds `Remove-TeamViewerCompany` to delete the company / tenant (Cannot be reverted, use with care!).